Refactored getProviders, fixed testing, and corrected godoc.
found that getProviders would be useful to have elsewhere so
we made it so. Godoc convention was not followed prior, but this
commit corrects that. Testing was very much hard-coded to use
the Linux on Mac environment, but has been changed to be more
flexible in which environments the tests run. Small, redundant
typecasting removed. Gofmt applied to files previously without.